se agrego el calculo de las rutas

+++ app/src/main/java/com/example/tesis/core/Individual.kt
@@ -2,7 +2,7 @@ package com.example.tesis.core
 
 data class Individual (
     var list: MutableList<Int> = mutableListOf(),
-    var distance: Float = 0f
+    var distance: Double = 0.0
 ) : Comparable<Individual> {
     override fun compareTo(other: Individual): Int {
         return distance.compareTo(other.distance)
